proycer/requirements.txt
Django==1.8.6
django-appconf==1.0.1
django-geoposition==0.2.2
django-smart-selects==1.1.1
Pillow==3.0.0
six==1.10.0
wheel==0.24.0
